fix(template) Proper parse exception for inconsistant variables
