+++ packages/markdown-template/lib/errorutil.js
@@ -14,7 +14,79 @@
 
 'use strict';
 
+const ParseException = require('@accordproject/concerto-core').ParseException;
 const TemplateException = require('./templateexception');
+
+/**
+ * Minimum length of expected token
+ * @param {object} expected the expected token
+ * @return {number} the minimum length
+ */
+function maxOfExpected(expected) {
+    return Math.max.apply(null,expected.map((x) => x.length));
+}
+
+/**
+ * Clean up expected tokens
+ * @param {object} expected the expected token
+ * @return {object} nicer looking expected tokens
+ */
+function cleanExpected(expected) {
+    return expected.map((x) => new RegExp(/'[^']*'/).test(x) ? x.substr(1,x.length -2) : x);
+}
+
+/**
+ * Throw a parse exception
+ * @param {string} markdown a markdown string
+ * @param {object} result the parsing failure
+ * @param {string} [fileName] - the fileName for the markdown (optional)
+ */
+function _throwParseException(markdown,result,fileName) {
+    // File location
+    const fileLocation = {};
+    let shortMessage;
+    let longMessage;
+    if (typeof result !== 'string') {
+        // Short message
+        shortMessage = `Parse error at line ${result.index.line} column ${result.index.column}`;
+
+        // Location
+        const start = result.index;
+        const end = Object.assign({},start);
+        end.offset = end.offset+1;
+        end.column = end.column+1;
+        fileLocation.start = start;
+        fileLocation.end = end;
+
+        const lines = markdown.split('\n');
+        const expected = result.expected;
+        const underline = ((line) => {
+            const maxLength = line.length - (start.column-1);
+            const maxExpected = maxOfExpected(cleanExpected(expected));
+            return '^'.repeat(maxLength < maxExpected ? maxLength : maxExpected);
+        });
+        const line = lines[start.line - 1];
+        const snippet = line + '\n' + ' '.repeat(start.column-1) + underline(line);
+        const isEOF = (x) => {
+            if (x[0] && x[0] === 'EOF') {
+                return true;
+            } else {
+                return false;
+            }
+        };
+
+        // Long message
+        const expectedMessage = 'Expected: ' + (isEOF(expected) ? 'End of text' : expected.join(' or '));
+        longMessage = shortMessage + '\n' + snippet + '\n' + expectedMessage;
+    } else {
+        shortMessage = result;
+        longMessage = shortMessage;
+        fileLocation.start = { offset: -1, column: -1 };
+        fileLocation.end = { offset: -1, column: -1 };
+    }
+
+    throw new ParseException(shortMessage, fileLocation, fileName, longMessage, 'markdown-template');
+}
 
 /**
  * Throw a template exception for the element
@@ -47,3 +119,4 @@ function _throwTemplateExceptionForElement(message, element) {
 }
 
 module.exports._throwTemplateExceptionForElement = _throwTemplateExceptionForElement;
+module.exports._throwParseException = _throwParseException;
